Automatically Generated Documentation

DEXTra.DEXGraph — Type

DEXGraph

A graph representation of a set of DEXs. Although the

Fields

graph: The name of the blockchain this liquidity pool is on.

nodes: A list of the unique node labels.

positions: The position of a particular node in nodes list.

DEXTra.LiquidityPool — Type

LiquidityPool

A liquidity pool represents a single source of assets and can be thought of as the nodes in a graph. All trading must occur between liquidity pools.

Fields

chain: The name of the blockchain this liquidity pool is on.

coin: The name of the coin

liquidity: The amount of liquidity in the pool denominated it's own coin.

name: Allows addtional identification options if there are multiple liquidity pools for the same coin on the same chain.

DEXTra.Order — Type

Order

A tuple representing an order in the orderbook.

Fields

β: The index of the asset being bought.

σ: The index of the asset being sold.

x̄: The maximum buy amount.

ȳ: The maximum sell amount.

π: The maximum exchange rate (i.e. pβ/pσ ≤ π).

DEXTra.OrderBook — Type

OrderBook

A stack of orders.

Fields

orders: A vector of orders.

assets: A vector of asset names.

DEXTra.TradingPair — Type

TradingPair

A trading pair represents a pair of liquidity pools and the exchange function between these pools.

Note

The two pools should always be ordered alphabetically.

Fields

provider: The name of the service provider ex. Uniswap, Sushiswap, etc.

pool1: The first liquidity pool in the trading pair.

pool2: The second liquidity pool in the trading pair.

fee: The fee charged for exchanges on this trading pair.

exchange_func: The exchange function between the two pools.

DEXTra.LimitBuyOrder — Method

LimitBuyOrder

A tuple representing a limit buy order in the orderbook.

Fields

β: The index of the asset being bought.

σ: The index of the asset being sold.

x̄: The maximum buy amount.

π: The maximum exchange rate (i.e. price[β]/price[σ] ≤ π).

DEXTra.LimitSellOrder — Method

LimitSellOrder

A tuple representing a limit sell order in the orderbook.

Fields

β: The index of the asset being bought.

σ: The index of the asset being sold.

ȳ: The maximum sell amount.

π: The maximum exchange rate (i.e. price[β]/price[σ] ≤ π).

DEXTra.MarketBuyOrder — Method

MarketBuyOrder

A tuple representing a market buy order in the orderbook.

Fields

β: The index of the asset being bought.

σ: The index of the asset being sold.

x̄: The maximum buy amount.

DEXTra.MarketSellOrder — Method

MarketSellOrder

A tuple representing a market sell order in the orderbook.

Fields

β: The index of the asset being bought.

σ: The index of the asset being sold.

ȳ: The maximum sell amount.

DEXTra.construct_model — Method

construct_model(opt_func::Function, orderbook::OrderBook)

Construct a JuMP model for the CoW problem.

Arguments

opt_func: The function to optimize

orderbook: The orderbook.

DEXTra.get_asset_map — Method

get_asset_map(orderbook::OrderBook)

Get a dictionary that maps the index of the assets in the orderbook to the index of the assets in the prices vector. (i.e. if 3 => 1 is in the pmap then prices[1] will be used to represent the price of asset 3). This allows you have a small vector representation of the prices.

Arguments

orderbook: The orderbook.

DEXTra.get_graph — Method

get_graph(orderbook::OrderBook, prices::Vector; min_val=0, max_val=1000)

Plot the orderbook as a graph.

Arguments

orderbook: The orderbook.

prices: A vector of representing all the prices.

DEXTra.get_level_graph — Method

get_level_graph(dex_graph::DEXGraph, source::String, sink::String)

Obtain a level graph from a DEXGraph object

DEXTra.get_univariate — Method

get_univariate(tps::Array{TradingPair}; directions::Array{Symbol})::Function

Compose multiple univariate exchange functions into a single function.

DEXTra.get_univariate — Method

get_univariate(tp::TradingPair; direction::Symbol)::Function

Return the univariate exchange function for a given trading pair, using the current liquidity values.

DEXTra.total_trade_volume — Method

total_trade_volume(orderbook::Orderbook, buy_volumes::Vector{Float64}, sell_volumes::Vector{Float64}, prices::Vector{Float64})

Compute the total trade volume of an orderbook given the buy and sell volumes and prices of each asset.

Arguments

orderbook: The orderbook.

buy_volumes: A vector of buy volumes for each asset.

sell_volumes: A vector of sell volumes for each asset.

prices: A vector of representing all the prices.

Kwargs

pmap: A dictionary that maps the index of the asset prices (i.e. if 3 => 1 is in the pmap then prices[1] will be used to represent the price of asset 3). This allows you have a small vector representation of the prices.

Graphs.SimpleGraphs.rem_vertex! — Method

rem_vertex!(dex_graph::DEXGraph, vertex::String)

Remove a vertex by the name of the node and re-index the nodes and positions fields.

DEXTra.ExchangeFunctions.constant_product — Method

constant_product(δ::Real, γ::Real, R₁::Real, R₂::Real)::Real

The exchange function for a constant-product market maker (CPMM)