Automatically Generated Documentation

DEXTra.DEXGraph — Type

DEXGraph

A graph representation of a set of DEXs. Although the

Fields

graph: The name of the blockchain this liquidity pool is on.

nodes: A list of the unique node labels.

positions: The position of a particular node in nodes list.

DEXTra.LiquidityPool — Type

LiquidityPool

A liquidity pool represents a single source of assets and can be thought of as the nodes in a graph. All trading must occur between liquidity pools.

Fields

chain: The name of the blockchain this liquidity pool is on.

coin: The name of the coin

liquidity: The amount of liquidity in the pool denominated it's own coin.

name: Allows addtional identification options if there are multiple liquidity pools for the same coin on the same chain.

DEXTra.TradingPair — Type

TradingPair

A trading pair represents a pair of liquidity pools and the exchange function between these pools.

Note

The two pools should always be ordered alphabetically.

Fields

provider: The name of the service provider ex. Uniswap, Sushiswap, etc.

pool1: The first liquidity pool in the trading pair.

pool2: The second liquidity pool in the trading pair.

fee: The fee charged for exchanges on this trading pair.

exchange_func: The exchange function between the two pools.

DEXTra.get_level_graph — Method

get_level_graph(dex_graph::DEXGraph, source::String, sink::String)

Obtain a level graph from a DEXGraph object

DEXTra.get_univariate — Method

get_univariate(tps::Array{TradingPair}; directions::Array{Symbol})::Function

Compose multiple univariate exchange functions into a single function.

DEXTra.get_univariate — Method

get_univariate(tp::TradingPair; direction::Symbol)::Function

Return the univariate exchange function for a given trading pair, using the current liquidity values.

Graphs.SimpleGraphs.rem_vertex! — Method

rem_vertex!(dex_graph::DEXGraph, vertex::String)

Remove a vertex by the name of the node and re-index the nodes and positions fields.

DEXTra.ExchangeFunctions.constant_product — Method

constant_product(δ::Real, γ::Real, R₁::Real, R₂::Real)::Real

The exchange function for a constant-product market maker (CPMM)